Added the feature "search most recent commit". Now, by specifying the "-r" flag, the commits that will be returned, are only the most recent one for every file that has changed for every commit. Changed slightly the order of the functions to be more precise and deleted some useless parameter from function invocation. Deleted the previous main.py file and renamed the new one.
